move utility files to utilities folder
