+++ src/utilities/animation-utility.js
